* Feature: Support definition of custom attributes on agent registration
Agent will read a JSON encoded array of key/value pairs from the
ECS_INSTANCE_ATTRIBUTES environment variable or via the specified JSON
configuration file. Instance attributes are passed to the ECS service during
RegisterContainerInstance() and are applied to the container instance in a
manner similar to that defined for the putAttribute API call and may be used to
influence Task Placement constraints.

Attribute registration is only performed on initial registration of the agent
with a cluster.

If the contents of the ECS_INSTANCE_ATTRIBUTES environment variable are
unparseable as JSON, the agent will exit with a terminal status code (5)
without attempting to register with the cluster. If variable contains valid
JSON, but the values do not comply with the ECS constraints around instance
attributes, the agent will complete registration with the cluster, but will
disconnect and terminate with exit code 5.

Support custom user-defined attributes when registering the agent.

Update TestEnvironmentConfig unit tests to use the testify library.

Pass custom attributes to RegisterContainerInstance only on new registrations

Update client tests to cover custom attributes and new instance registration

Handle service responses to invalid instance registration attributes

Add unit test for validateRegisteredAttributes() and incorporate PR feedback:
 * Incorporate PR feedback from aaithal around formatting and structure.
 * Incorporate PR feedback: use seelog.
 * PR feedback from richardpen:
 * ensure that tests cover attributes with name but no values.
 * readability improvements